increase the number of things to get per iteration
place_holder does not do what I thought it would. If there have been
more than limit things since the last update, the rest won't be recorded
in the next one.

This probably wouldn't be a problem even with a low limit because the
plounge is so low traffic but I am paranoid so I increased the limit to
100 (the most possible with one request).

using reddits 'before' parameter would probably be a better solution but
I don't know how it reacts if things get deleted. Sorry for the
excessive load reddit :(.